Break up ChartView into smaller components
