+++ components/ChartView.tsx
@@ -1,85 +1,26 @@
 import { mapRef } from "@/hooks/useMapRef";
-import { loadMarkers } from "@/hooks/useMarkers";
 import { addDraftPoint, insertDraftPointAt, useRouteDraft } from "@/hooks/useRouteDraft";
-import { useSelection, useSelectionHandler } from "@/hooks/useSelection";
-import { useSheetOffset } from "@/hooks/useSheetPosition";
-import useTheme from "@/hooks/useTheme";
-import { mapStyles, useViewOptions } from "@/hooks/useViewOptions";
-import { getDistance } from "geolib";
+import { useSelectionHandler } from "@/hooks/useSelection";
+import { useMapStyle } from "@/hooks/useViewOptions";
+import { findNearestLegIndex } from "@/lib/geo";
 import { Images, Map } from "@maplibre/maplibre-react-native";
-import { router } from "expo-router";
-import { Fragment, useCallback, useEffect } from "react";
-import Animated from "react-native-reanimated";
-import { SafeAreaView } from "react-native-safe-area-context";
+import { useCallback } from "react";
 import AISLayer from "./AISLayer";
 import AtoNLayer from "./AtoNLayer";
-import { MapControls } from "./MapControls";
+import MapOverlay from "./MapOverlay";
 import MarkerOverlay from "./MarkerOverlay";
-import NavigationHUD from "./NavigationHUD";
 import RouteOverlay from "./RouteOverlay";
 import TrackOverlay from "./TrackOverlay";
-import { Annotation } from "./map/Annotation";
 import { NavigationCamera, handleRegionDidChange, handleRegionIsChanging } from "./map/NavigationCamera";
 import { NavigationPuck } from "./map/NavigationPuck";
-import TrackRecordButton from "./map/TrackRecordButton";
-
-/**
- * Find the index of the leg segment closest to a point.
- * Returns the index to insert at (i.e. after points[index-1], before points[index]),
- * or null if no leg is within the threshold.
- */
-function findNearestLegIndex(
-  lat: number,
-  lon: number,
-  points: { latitude: number; longitude: number }[],
-  thresholdMeters: number,
-): number | null {
-  if (points.length < 2) return null;
-
-  let bestDist = Infinity;
-  let bestIndex: number | null = null;
-
-  for (let i = 0; i < points.length - 1; i++) {
-    const a = points[i];
-    const b = points[i + 1];
-    // Project point onto segment using simple lat/lon interpolation
-    const dx = b.longitude - a.longitude;
-    const dy = b.latitude - a.latitude;
-    const lenSq = dx * dx + dy * dy;
-    if (lenSq === 0) continue;
-    const t = Math.max(0, Math.min(1, ((lon - a.longitude) * dx + (lat - a.latitude) * dy) / lenSq));
-    const projLat = a.latitude + t * dy;
-    const projLon = a.longitude + t * dx;
-    const dist = getDistance({ latitude: lat, longitude: lon }, { latitude: projLat, longitude: projLon });
-    if (dist < bestDist) {
-      bestDist = dist;
-      bestIndex = i + 1; // insert after point i
-    }
-  }
-
-  return bestDist <= thresholdMeters ? bestIndex : null;
-}
+import SelectedLocationAnnotation from "./map/SelectedLocationAnnotation";
 
 export default function ChartView() {
-  const mapStyleId = useViewOptions((s) => s.mapStyleId);
-  const selection = useSelection();
-  const sheetOffset = useSheetOffset();
-  const theme = useTheme();
-  const mapStyle = mapStyles.find(style => style.id === mapStyleId)?.style || mapStyles[0].style;
-
-  useEffect(() => {
-    loadMarkers();
-  }, []);
-
+  const mapStyle = useMapStyle();
   const navigate = useSelectionHandler();
-  const draftActive = useRouteDraft((s) => s.points.length > 0);
-  const selectedCoords = selection?.type === "location"
-    ? selection.id.split(",").map(Number) as [number, number]
-    : null;
 
   const handleLongPress = useCallback((e: { nativeEvent: { lngLat: [number, number] } }) => {
     const points = useRouteDraft.getState().points;
-    if (points.length === 0) return;
 
     const [lon, lat] = e.nativeEvent.lngLat;
 
@@ -93,12 +34,6 @@ export default function ChartView() {
     }
   }, []);
 
-  const handleDragEnd = useCallback(
-    (e: { nativeEvent: { lngLat: [number, number] } }) =>
-      router.setParams({ id: `${e.nativeEvent.lngLat[0]},${e.nativeEvent.lngLat[1]}` }),
-    []
-  );
-
   return <>
     <Map
       ref={mapRef}
@@ -111,25 +46,22 @@ export default function ChartView() {
       compassPosition={{ top: -2000, right: -2000 }}
       onRegionIsChanging={handleRegionIsChanging}
       onRegionDidChange={handleRegionDidChange}
-      onLongPress={draftActive ? handleLongPress : undefined}
+      onLongPress={handleLongPress}
       onPress={(e) => {
         const { lngLat } = e.nativeEvent;
-        navigate("location", `${lngLat[0]},${lngLat[1]}`);
+
+        const { points } = useRouteDraft.getState();
+        if (points.length > 0) {
+          // const [lon, lat] = lngLat;
+          // addDraftPoint({ latitude: lat, longitude: lon });
+          return;
+        }
+
+        navigate("location", lngLat.join(','));
       }}
       logo={false}
     >
       <NavigationCamera />
-      {selectedCoords && (
-        <Annotation
-          id="selected-location"
-          lngLat={selectedCoords}
-          icon="mappin"
-          color={theme.danger}
-          selected
-          draggable
-          onDragEnd={handleDragEnd}
-        />
-      )}
       <Images images={{
         "vessel-default": { source: require("@/assets/vessels/png/default.png"), sdf: true },
         "vessel-unknown": { source: require("@/assets/vessels/png/unknown.png"), sdf: true },
@@ -148,25 +80,14 @@ export default function ChartView() {
         "aton-lighthouse": { source: require("@/assets/atons/png/lighthouse.png"), sdf: true },
         "aton-virtual": { source: require("@/assets/atons/png/virtual.png"), sdf: true },
       }} />
-      <Fragment key={mapStyleId}>
-        <TrackOverlay />
-        <MarkerOverlay />
-        <RouteOverlay />
-        <AISLayer />
-        <AtoNLayer />
-        <NavigationPuck />
-      </Fragment>
+      <TrackOverlay />
+      <MarkerOverlay />
+      <RouteOverlay />
+      <AISLayer />
+      <AtoNLayer />
+      <SelectedLocationAnnotation />
+      <NavigationPuck />
     </Map>
-    <SafeAreaView style={{ position: "absolute", top: 0, left: 16, right: 16, alignItems: "center" }}>
-      <NavigationHUD />
-    </SafeAreaView>
-    <Animated.View style={[{ position: "absolute", bottom: 0, left: 0, right: 0 }, sheetOffset]}>
-      <SafeAreaView style={{ position: "absolute", bottom: 0, right: 16, gap: 16 }}>
-        <MapControls />
-      </SafeAreaView>
-      <SafeAreaView style={{ position: "absolute", bottom: 0, left: 16, gap: 16 }}>
-        <TrackRecordButton />
-      </SafeAreaView>
-    </Animated.View>
+    <MapOverlay />
   </>;
 }
