+++ components/MapOverlay.tsx
@@ -1,0 +1,26 @@
+import { useSheetOffset } from "@/hooks/useSheetPosition";
+import Animated from "react-native-reanimated";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { MapControls } from "./MapControls";
+import NavigationHUD from "./NavigationHUD";
+import TrackRecordButton from "./map/TrackRecordButton";
+
+export default function MapOverlay() {
+  const sheetOffset = useSheetOffset();
+
+  return (
+    <>
+      <SafeAreaView style={{ position: "absolute", top: 0, left: 16, right: 16, alignItems: "center" }}>
+        <NavigationHUD />
+      </SafeAreaView>
+      <Animated.View style={[{ position: "absolute", bottom: 0, left: 0, right: 0 }, sheetOffset]}>
+        <SafeAreaView style={{ position: "absolute", bottom: 0, right: 16, gap: 16 }}>
+          <MapControls />
+        </SafeAreaView>
+        <SafeAreaView style={{ position: "absolute", bottom: 0, left: 16, gap: 16 }}>
+          <TrackRecordButton />
+        </SafeAreaView>
+      </Animated.View>
+    </>
+  );
+}
